requirements.txt
pillow==5.3.0
pygame==1.9.4
flake8==3.7.5
pytest==3.10.1
